Add decimal-maximum-digits attr to number-text column (#1611)
## 🤨 Rationale

This feature is needed to eventually render file sizes in the Nimble
table in a way consistent with how SLE renders them today. Specifically,
it renders them with one decimal place, except if it would be a zero, in
which case it renders without any decimal places.

## 👩‍💻 Implementation

The `Intl.NumberFormat` used by the number-text column has options
`maximumFractionDigits` and `minimumFractionDigits`. We had always been
setting both options to the value of `decimal-digits` set by the client.
Now, if `decimal-maximum-digits` is set rather than `decimal-digits`, we
use that for `maximumFractionDigits` and pass 0 for
`minimumFractionDigits`.

Additionally:

- The default alignment of a `decimal`-formatted column is not always
`right` anymore. If `decimal-maximum-digits` is set, then the default
alignment is `left`, because decimal points are no longer guaranteed to
align.
- `decimal-maximum-digits` is mutually exclusive with `decimal-digits`,
and a validation error is set if both are set.
- A validation error is set if the value of `decimal-maximum-digits` is
set outside the range of 0-20.
- Updated Storybook documentation
- Added unit tests
- Updated Blazor wrapper and added one test
- Updated Angular directive and added tests

## 🧪 Testing

Unit tests pass

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.