Update tanstack dependencies (#1621)
# Pull Request

## 🤨 Rationale

An issue was discovered when trying to automatically update the
dependency on TanStack. This PR updates those dependencies and fixes the
issue on the label-provider storybook page.

Resolves #1617

## 👩‍💻 Implementation

- Update version of `@tanstack/table-core` and `@tanstack/virtual-core`
- Set a `min-height` on the tables used in the label-provider stories

## 🧪 Testing

Verified the tests pass and there are no chromatic changes

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>