Configure Renovate to rebase PRs less frequently (#1640)
# Pull Request

## 🤨 Rationale

Once Renovate has created PRs it currently updates them any time
something else changes in the repo. This is annoying because it uses
lots of build resources and makes it hard to track down a completed
build for the PR, for example to inspect Storybook.

In [a PR
discussion](https://github.com/ni/nimble/pull/1634#discussion_r1379067152)
we decided to switch to manually triggering the rebase before merging
the PR.

## 👩‍💻 Implementation

1. Update renovate.json to `rebaseWhen: conflicts`. This seems more
helpful than `never` but hopefully less noisy than `auto`.
2. Update CONTRIBUTING to mention that owners may need to manually
trigger the rebase


## 🧪 Testing

We'll see if it works the next time Renovate creates PRs.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.