Update fast-foundation dependency (#1642)
# Pull Request

## 🤨 Rationale

The latest version of fast-foundation has fixes for two of our toolbar
issues, so this PR pulls in the latest version.

Resolves #1422 
Resolves #1423 

## 👩‍💻 Implementation

Bump version in package file
Update the rich-text-editor page object to adjust for a change in the
toolbar code

## 🧪 Testing

Used storybook to verify that the two toolbar issues are resolved

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.