Expose width and height of nimble-dialog & add tokens (#1553)
# Pull Request

## 🤨 Rationale

<!---
Provide some background and a description of your work.
What problem does this change solve?

Include links to issues, work items, or other discussions.
-->

#1117

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as
complexity, alternative solutions, performance impact, etc.

Consider listing files with important changes or comment on them
directly in the pull request.
-->

- Performed the tasks that are enumerated in the issue linked above.
- Create 2 tokens for the existing (small) dialog: width & max-height
- Create 2 tokens for the large dialog: width & height
- Have the small dialog as the default (width & max-height of small
dialog; height set to `fit-content`, this was the default height that
was applied to the dialog until now)
- Added storybook documentation.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->
Manually thorugh the storybook tests & matrix tests in chromatic. The
sizes are configurable in the storybook similarly to how the width is
configurable for the nimble drawer. For the matrix tests, I added 2
stories: one for the small dialog and one for the large dialog:
https://www.chromatic.com/build?appId=60e89457a987cf003efc0a5b&number=7776


## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [X] I have updated the project documentation to reflect my changes or
determined no changes are needed.