Rich text editor | Nimble listbox component bring up for mention popup (#1654)
# Pull Request

## 🤨 Rationale

<!---
Provide some background and a description of your work.
What problem does this change solve?

Include links to issues, work items, or other discussions.
-->

The `nimble-rich-text-mention-list-box` is the component used to
populate the list of mention options in the editor. To enable some of
the interactions within the popup, it requires FAST's listbox component.
As the listbox has not been created in Nimble yet, this PR introduces a
new `nimble-listbox` component.

For more details about mention list box component, see
[mention-hld](https://github.com/ni/nimble/blob/main/packages/nimble-components/src/rich-text/specs/mention-hld.md#mention-popup-visible-ui-element).

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as
complexity, alternative solutions, performance impact, etc.

Consider listing files with important changes or comment on them
directly in the pull request.
-->

1. Create a new `nimble-listbox` component from `fast-listbox`
2. Add to the `Tests/` section in the storybook as it is not intended
for end users usage and it is for the implementation of
`nimble-rich-text-mention-list-box` for mention popup.
3. Create `menuMinWidth` token to share the min-width between `menu` and
`listbox`

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->

1. Basic test cases are written for the new `nimble-listbox` component
2. Storybook matrix test

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.