Fix the drawer's footer's border to match the visual design spec (#1672)
# Pull Request

## 🤨 Rationale

The top border of the drawer's footer should be 2px, rather than 1px. It
was also not quite the right color. I've updated the border to be styled
in an identical fashion to the similar border at the top of the dialog's
footer, which is correct.

## 👩‍💻 Implementation

- Minor style update

## 🧪 Testing

- Verified the styling looks correct in storybook

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.