Rich Text Editor | Add Configuration component for @mention support (#1652)
# Pull Request

## 🤨 Rationale
The list of items needs to be populated in the mention popup of the
editor when entering configured characters (mentioning feature), in our
case user list need to be populated when entering `@` character in the
editor. To get the configuration of the users from the client and pass
on to the editor, Configuration components are needed.

Link to the Feature: [Feature
2351134](https://dev.azure.com/ni/DevCentral/_workitems/edit/2351134):
Support user @ mention and send a plain text email notification

## 👩‍💻 Implementation

- Added user mapping component `nimble-mapping-user` by extending the
mapping base class. Added attribute `display-name` for textual
representation of user name.
- Added `RichTextMention` abstract class, which will hold
- `mentionInternals` contains mention configuration, icon, character and
pattern data which will be used in the editor.
- `checkValidity()` and `validity` getter - will return the validity of
the configuration.
         - Generated configuration from mapping element

- Added `MentionValidator` and `MentionInternals` model classes.
- Added `RichTextMentionBaseValidator` base validator to check the below
cases
          - unsupportedMappingType
          - duplicateMappingMentionHref
          - missingMentionHrefValue
          - unsupportedMentionHrefValue
          - missingDisplayNameValue
          - missingPatternAttribute
          - unsupportedPatternValue

- Added `nimble-rich-text-mention-users` component derived from the base
class which will hold definitions for the abstract methods to configure
user mentioning in editor.

## 🧪 Testing

- Added unit tests for the component.
- Added basic storybook doc, will be linked to editor and viewer mdx
files once created.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: rajsite <rajsite@users.noreply.github.com>