Rich text editor | Update testing strategy in markdown-serializer (#1633)
# Pull Request

## 🤨 Rationale
Testing strategy in the markdown serializer need to be updated as it is
generating a mock editor currently to test the behavior instead of
implementing actual editor. So, it needs to be updated with actual
editor to test the production configuration.

Issue link: #1533 

## 👩‍💻 Implementation

- Removed the mock editor and constructed the actual editor
`nimble-rich-text-editor` to get the editor document.
- Removed all the html as input and updated it with `setMarkdown()`
method as in production flow only there will be markdown interaction
with the editor and not with the html.

## 🧪 Testing

- Manually ran the test and verified.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.