Hide inc/dec buttons on numeric from a11y tree (#1677)
## 🤨 Rationale

Fixes #1644

## 👩‍💻 Implementation

Added `aria-hidden="true"` to both inc/dec buttons. As these buttons are
not focusable, and there are other keyboard interactions (up/down arrow)
for incrementing/decrementing the value, these buttons are not
interesting to assistive technologies. The `<button>` element implicitly
has `focusable=true` for a11y, and this cannot be changed, even if it is
removed from the tab order with `tabIndex=-1`. If we want to keep it as
an unfocusable button, we must use `aria-hidden`.

## 🧪 Testing

Checked in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.