Handle overflow on nimble-list-option elements (#1669)
# Pull Request

## 🤨 Rationale

Resolves #1658 

I also noticed that we had a workaround in the nimble-list-option for a
FAST issue that has been fixed upstream, so I removed the workaround and
validated that the issue is in fact fixed.

## 👩‍💻 Implementation

- Added styling to the nimble-list-option to ellipsis the content if it
gets too long
- Leverage existing `overflow` directive to the list option template so
that a `title` will be applied when the text is too long
- This required forking the FAST template for the list option, so in
addition to forking the template, I also brought in the FAST tests.

## 🧪 Testing

- Manually tested overflow behavior in storybook
- Manually tested that issue that had a workaround is still behaving as
expected with the workaround removed.
- Wrote new unit tests to verify the overflow/title behavior
- Brought in the FAST unit tests for the list option since the template
has now been forked

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.