Rich text | Support for @mention in Editor (#1683)
# Pull Request

## 🤨 Rationale

[User Story
2536827](https://dev.azure.com/ni/DevCentral/_workitems/edit/2536827):
FE | Add @ mention support in Nimble rich-text-editor and in markdown
serializer


## 👩‍💻 Implementation

- Added `@tiptap/extension-mention` extension for mention support and
configured for user mention support.
- Added `mention` event which will be emitted from the respective
`RichTextMention` configuration element when the user is typing `@` in
the editor.
- Added support for `RichTextMentionUsers` configuration element in the
editor.
- Added `getMentionedHrefs()` method in the configuration element for
the client to get the mention list id (`href`) which already exists in
the editor/viewer
   
- In editor the `hrefs` will be generated from the
`RichTextMarkdownSerializer` as it gets dynamically updated.
- In the viewer from `RichTextMarkdownParser`, 'hrefs' will be generated
as it will be updated while parsing.

- Configured mention tag and decoration tag (Tag added when user typing
`@` ) as `nimble-rich-text-mention-users-view` element for mention Node.
- Created abstract base class `RichText` for rich text components
(editor and viewer). Common methods have been relocated to this base
class, which is then extended by the editor and viewer classes. (As the
viewer and editor have the common functionality for getting the
configuration from `RichTextMentionUsers` configuration elements, moved
it to a base class).
- Added `MentionNode` in `RichTextMarkdownSerializer` to serialize the
mentioned node. (markdown format is same as of auto-link)
- Update markdown-parser to add support for mention node in the editor
- Handled copy-paste functionality for mention in the editor. Currently,
when mentioned node is copied and pasted into the editor it will render
it as plain text. For example, copying `@user` and pasting it into the
editor will render it as `user`. The `@` character is stripped off as
the text node gets generated only based on the `label` attribute of
`rich-text-mention-users-view` while transform pasting. Since in the
user view element, the `@` character is added on the template view, the
label is missing the character.
- Modified the serializer from being static to having an instance. This
change enables dynamic initialization of the serializer when new
mentions are added, allowing for the serialization of the respective
mention nodes.


## 🧪 Testing

- Added tests to check the dynamic loading of child elements into the
editor component
- Added tests for parser and serializer `getMentionedHrefs` logic to
check whether it is returning the existing mention `href` list.
- Added storybook doc (`rich-text-editor.mdx`) and linked mention
configuration docs as well.
- Added tests for checking the copy/paste behavior of mention nodes.

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: rajsite <rajsite@users.noreply.github.com>
Co-authored-by: Vivin Krishna <123377523+vivinkrishna-ni@users.noreply.github.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>