Rich text editor | Fix mention listbox opening at the wrong position when mention dynamically loads (#1709)
# Pull Request

## 🤨 Rationale

<!---
Provide some background and a description of your work.
What problem does this change solve?

Include links to issues, work items, or other discussions.
-->

Whenever the `@mention` is triggered, moving the cursor away from the
decoration element and updating the configuration element dynamically
opens the popup at the top right of the page.

In the GIF below, we are updating the configuration element by updating
the `Load Config` button to demonstrate the issue

![popup-at-the-top-issue-nimble](https://github.com/ni/nimble/assets/123377523/4a81a93e-8dec-41a7-906c-7da494572227)

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as
complexity, alternative solutions, performance impact, etc.

Consider listing files with important changes or comment on them
directly in the pull request.
-->

The issue is because the `onStart()` event of the suggestion plugin of
Tiptap is triggered whenever we perform `setMarkdown()` when the
configuration element updates dynamically. However, if we move the
cursor to start as mentioned in the above GIF which triggers the
`onExit()` event and then `onStart()` the props in `onStart()` called
without a decoration node to load the popup right below `@` character.

So in `onstart()`, if the decoration node is null, we restrict opening
the popup.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->
- Written a test for the popup issue with the same workflow.
- Manually verified in SLE for the same popup issue

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.