Fix for Safari (WebKit) browser issue in removing the suggestion text when mentioning a user (#1721)
# Pull Request

## 🤨 Rationale

Fixes: [Bug
2606385](https://dev.azure.com/ni/DevCentral/_workitems/edit/2606385):
Mention | Removing the suggestion character removes the whole word in
iOS and Mac

## 👩‍💻 Implementation

This PR includes rendering the default HTML span element for the
suggestion node by Tiptap, instead of a view element during editing
`@mention`. This is achieved by removing the decoration tag
configuration in the Mention extension.

Our idea is to use a view element in two states: edit mode (during the
mention popup is opened) and view mode (after committing from mention
popup). See here for more details on the design discussion:
https://github.com/ni/nimble/pull/1653#discussion_r1385849063. However,
we have an issue in Safari browsers on Mac and all browsers on iOS &
iPadOS where if the user presses backspace, it removes the whole
suggestion text. For example, if I type `@john` and press backspace
should delete a single character (`@joh`), but it deletes the whole
node.

This is because whenever we render a shadow root element (View element)
within a contenteditable div like editor, pressing backspace removes the
whole content in Safari whereas, it removes a single character in other
browsers. Here is the stackblitz example to reproduce:
https://stackblitz.com/edit/typescript-avgrtp?file=package.json%2cindex.html

So I have created an issue to WebKit and here is the link to it:
https://bugs.webkit.org/show_bug.cgi?id=266700
In review we decided that we will keep the change of using a single span
style for all mention editing and only use custom mention views for
committed mentions. That change will happen in follow-up PR:
https://github.com/ni/nimble/pull/1731

## 🧪 Testing


Updated test scripts to check if `span` element is rendered when
`@mention` is triggered.
New tests were not added to validate backspace behavior. This should be
revisited as an integration test, captured here:
https://github.com/ni/nimble/issues/1568#issuecomment-1877589782

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>