Update parameterized tests (#1727)
# Pull Request

## 🤨 Rationale

This PR addresses the first part of #1551. It migrates tests in
`nimble-components` to the `parameterizeNamedList` utility and deletes
the deprecated `getSpecTypeByNamedList`.

To complete #1551, `parameterizeNamedList` needs to be exported and used
from `nimble-angular`, but this will be tackled in a separate PR.

## 👩‍💻 Implementation

- Use `parameterizeNamedList` in the places where
`getSpecTypeByNamedList` had been used (plus a few more places that
weren't using either)
- Delete `getSpecTypeByNamedList`

## 🧪 Testing

Ran unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>