Add tests for list-option, menu-item, tab, and tree-view Angular directives (#1733)
# Pull Request

## 🤨 Rationale

Fixes #318 

This PR adds tests for the list-option, menu-item, tab, and tree-view
Angular directives. These are the directives called out in #318 as
needing tests. The theme-provider was also mentioned in the issue, but
tests have already been added for that directive.

## 👩‍💻 Implementation

The bulk of the changes are adding nimble's standard Angular directive
tests for each directive that was missing tests.

The more interesting change is a new list-option directive to contain
the properties of the `nimble-list-option` component. It lives alongside
the select-list-option and combobox-list-option directives. However,
those directives previously each had their own duplicated implementation
for the `nimble-list-option` properties (currently only `disabled`).
This didn't seem ideal, and it also posed a problem with testing because
there wasn't a great way to test each directive in isolation because
they had the same selector, the same property, and were exported from
the same Angular module. Now, the list-option directive contains
everything that would be needed to use the list-option (exports types
and contains `disabled` property), and the combobox & select specific
directives for the list-option contain the form integration code that is
necessary to use the combobox and select within an Angular form. All
three directives are exported from the existing list-option module, so
this change should be unnoticed by clients.

## 🧪 Testing

Ran auto tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.