Rich Text Editor | Fix for losing the state of format buttons when the configuration is updated dynamically (#1729)
# Pull Request

## 🤨 Rationale

- Fix for [Bug
2614292](https://dev.azure.com/ni/DevCentral/_workitems/edit/2614292):
Editor format buttons active state (Bold and Italics) are not retained
when user configuration gets loaded dynamically

- Currently, when there are changes in the configuration, the states of
the bold and italics buttons are reverted to an inactive state. This is
happening because of the tiptap API
[setContent()](https://tiptap.dev/docs/editor/api/commands/set-content)
which will reset the formatting button states when setting content. This
PR will address this issue by resetting its states.

## 👩‍💻 Implementation

- Added the `resetEditorButtonsState()` method after `setMarkdown()`
method which will take previous button states as arguments and retain
the formatting button states.

## 🧪 Testing

- Written test cases to check the formatting button states on
configuration change.
- Manually tested and verified the functionality in Storybook.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.