Export tag name as literals (#1736)
# Pull Request

## 🤨 Rationale

We include type definitions such as the following so that TypeScript is
aware of the tag names we register:
```ts
declare global {
    interface HTMLElementTagNameMap {
        'nimble-anchor': Anchor;
    }
}
```

With that kind of definition using an API like
`document.createElement('nimble-anchor')` will let TypeScript infer the
result will be an instance of `Anchor`. However, because we were
exporting tag names as runtime calculated strings, TypeScript was not
able to infer the type using `document.createElement(anchorTag);`.

## 👩‍💻 Implementation

This change modifies our tag definitions to export them as constant
string literals to improve inferred typing. Notice how in some of the
tests `as` casts were removed by the linter for having unnecessary
casting.

## 🧪 Testing

Relied on exisiting tests.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed. Updated the element creation doc
snippets.