Remove `disable-editing` attribute from the mention view element (#1731)
# Pull Request

## 🤨 Rationale

<!---
Provide some background and a description of your work.
What problem does this change solve?

Include links to issues, work items, or other discussions.
-->

Previously, we used `nimble-rich-text-mention-users-view` for two states
one is at the time of triggering the mention (Edit mode) and the other
one is after committing the mention (View mode) in the editor. However,
due to https://github.com/ni/nimble/issues/1716, we wanted to remove the
decoration tag (which is configured in Tiptap to render the same element
in edit mode) here: https://github.com/ni/nimble/pull/1721. Also, during
editing for all mentions we can have the same styling so we decided not
to use this element for both edit and view mode. Only after committing
in the editor and for viewing in the viewer, this view element can be
used. During editing mention node, the styles can be explicitly
mentioned in the editor styles.

Relevant discussion:
https://github.com/ni/nimble/pull/1721#discussion_r1440837527

Will close the following issues upon merging this PR:
1. fixes https://github.com/ni/nimble/issues/1659
2. fixes https://github.com/ni/nimble/issues/1716

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as
complexity, alternative solutions, performance impact, etc.

Consider listing files with important changes or comment on them
directly in the pull request.
-->

- Removed `disable-editing` attribute from the view element base class
- Removed styles related to `disable-editing` and other CSS cleanups in
view element
- Added
[`decorationClass`](https://tiptap.dev/docs/editor/api/utilities/suggestion#decoration-class)
in the editor to style the default suggestion `span` in the editor and
added necessary styles to it.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->

- Removed the view element edit mode test case from the Chromatic matrix
test
- Manually tested the mention view element in editor and viewer

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>