Fix styling of disabled tree items (#1751)
# Pull Request

## 🤨 Rationale

Fixes #1358 

This PR fixes the styling of disabled tree items, both
`nimble-anchor-tree-item` and `nimble-tree-item`. There is an additional
oddity with disabled parent tree items in that they can be
expanded/collapsed with keyboard interactions but not with the mouse.
We've made the decision not to address (or track) this issue for now
because we don't have clear usage guidance of when it would make sense
to have a disabled parent tree item or what the behavior associated with
it should be.

## 👩‍💻 Implementation

The tree-item and anchor-tree-item disabled styles now match other
components (such as the list-option) and the Figma spec. The changes
made include:
- Making the cursor `default` rather than `not-allowed` when hovering
over a disabled tree item
- Making the hover color on a disabled tree item transparent
- Making the icon & text associated with a tree item
`bodyDisabledFontColor` rather than applying a 50% opacity. (Note that
`bodyDisabledFontColor` uses 30% opacity, so this did change the
rendered color).

## 🧪 Testing

- Manually tested in storybook
- Inspected the chromatic diffs

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.