Update Playwright version (#1728)
# Pull Request

## 🤨 Rationale

Fixes #1544 

## 👩‍💻 Implementation

- Added `postinstall` script to root package file that runs `playwright
install --with-deps` to ensure browsers get installed.
- Bumped dependency versions of Playwright to 1.40.0 in
nimble-components, nimble-blazor, and Angular workspace. They are now
pinned to one version, because the root workspace package file now
installs the browsers.
- Fixed two table page object functions that were incorrect.
- Added workaround (rounding cell width) for tests failing due to
pulling in new version of Chrome browser (from 115.0.5790.75 to
120.0.6099.28). Six table interactive column resize tests began failing
with errors like `Expected 100.984375 to be 101.` and `Expected
74.984375 to be 75.` I confirmed that this numerical imprecision is
related to using fractional width units (e.g. `1fr`).

We still cannot have Renovate update the Playwright dependency because
of the inconsistency of versions between the npm package and the nuget.
E.g. currently the latest npm package is 1.40.1, but the latest nuget is
only 1.40.0, so our build will fail as it tries to find a nuget with the
same version as the npm package.

## 🧪 Testing

Pipeline build succeeds

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>